+++ src/flexrml/backend/executor/complex_executor.cpp
@@ -14,6 +14,9 @@
 
 #include "csv_row.h"
 #include "definitions.h"
+#include "join_program.h"
+#include "term_cache.h"
+#include "term_program.h"
 #include "utils.h"
 #include "xxhash.h"
 
@@ -27,389 +30,6 @@ static void create_parent_directories_if_needed(const fs::path& path) {
   if (!parent.empty()) {
     fs::create_directories(parent);
   }
-}
-
-enum class CompiledTermMapType {
-  Preformatted,
-  Constant,
-  Reference,
-  Template
-};
-
-struct CompiledTemplatePart {
-  std::string literal;
-  int reference_index = -1;
-};
-
-struct CompiledTerm {
-  bool usable = false;
-  CompiledTermMapType map_type = CompiledTermMapType::Constant;
-  std::string term_map;
-  std::string term_type;
-  std::string lang_tag = "None";
-  std::string data_type = "None";
-  std::vector<CompiledTemplatePart> parts;
-  int reference_index = -1;
-  bool infer_datatype = false;
-  bool add_base_iri = false;
-};
-
-struct CompiledRuntimeTerm {
-  std::vector<std::string> content;
-  CompiledTerm compiled;
-  enum class RenderOp {
-    Null,
-    Function,
-    Preformatted,
-    Compiled,
-    Fallback
-  };
-  RenderOp render_op = RenderOp::Fallback;
-};
-
-struct ComplexTermCacheEntry {
-  std::string key;
-  std::string value;
-  std::string scratch;
-  bool computed = false;
-};
-
-struct CompiledComplexRenderPlan {
-  CompiledRuntimeTerm subject;
-  CompiledRuntimeTerm predicate;
-  CompiledRuntimeTerm object;
-  CompiledRuntimeTerm graph;
-  int subject_cache_id = -1;
-  int predicate_cache_id = -1;
-  int object_cache_id = -1;
-  int graph_cache_id = -1;
-  std::vector<ComplexTermCacheEntry> term_cache_entries;
-  bool has_graph = false;
-  bool needs_row_map = false;
-  bool has_object_condition = false;
-};
-
-///////////////////////////////////////////////////////////////////////////////////////////////////////77
-//// HELPER FUNCTIONS ////
-
-static int projected_column_index(const std::vector<std::string>& projected_header,
-                                  std::string_view name) {
-  for (std::size_t i = 0; i < projected_header.size(); ++i) {
-    if (projected_header[i] == name) {
-      return static_cast<int>(i);
-    }
-  }
-  return -1;
-}
-
-static bool is_dynamic_annotation(const std::string& annotation,
-                                  const std::vector<std::string>& projected_header) {
-  return annotation.starts_with("==FUNC==") ||
-         annotation.find('{') != std::string::npos ||
-         projected_column_index(projected_header, annotation) >= 0;
-}
-
-static bool compile_template_parts(const std::string& term_map,
-                                   const std::vector<std::string>& projected_header,
-                                   std::vector<CompiledTemplatePart>& parts) {
-  std::string literal;
-  literal.reserve(term_map.size());
-
-  for (std::size_t i = 0; i < term_map.size();) {
-    if (term_map[i] == '\\' && i + 1 < term_map.size() &&
-        (term_map[i + 1] == '{' || term_map[i + 1] == '}')) {
-      literal.push_back(term_map[i + 1]);
-      i += 2;
-      continue;
-    }
-
-    if (term_map[i] != '{') {
-      literal.push_back(term_map[i]);
-      ++i;
-      continue;
-    }
-
-    const std::size_t placeholder_start = i;
-    ++i;
-    std::string reference_name;
-    bool closed = false;
-    for (; i < term_map.size(); ++i) {
-      if (term_map[i] == '\\' && i + 1 < term_map.size() &&
-          (term_map[i + 1] == '{' || term_map[i + 1] == '}')) {
-        reference_name.push_back(term_map[i + 1]);
-        ++i;
-      } else if (term_map[i] == '}') {
-        closed = true;
-        ++i;
-        break;
-      } else {
-        reference_name.push_back(term_map[i]);
-      }
-    }
-
-    if (!closed) {
-      literal.append(term_map, placeholder_start, std::string::npos);
-      break;
-    }
-
-    if (!literal.empty()) {
-      parts.push_back(CompiledTemplatePart{std::move(literal), -1});
-      literal.clear();
-    }
-
-    const int index = projected_column_index(projected_header, reference_name);
-    if (index < 0) {
-      return false;
-    }
-    parts.push_back(CompiledTemplatePart{"", index});
-  }
-
-  if (!literal.empty()) {
-    parts.push_back(CompiledTemplatePart{std::move(literal), -1});
-  }
-  return true;
-}
-
-static CompiledTerm compile_term(const std::vector<std::string>& content,
-                                 const std::vector<std::string>& projected_header,
-                                 const std::string& base_uri) {
-  CompiledTerm term;
-  if (content.size() < 3 || content[1] == "function") {
-    return term;
-  }
-
-  term.term_map = content[0];
-  term.term_type = content[2];
-  const bool is_literal = term.term_type == "literal";
-  if (is_literal) {
-    if (content.size() > 3) {
-      term.lang_tag = content[3];
-    }
-    if (content.size() > 4) {
-      term.data_type = content[4];
-    }
-    if (is_dynamic_annotation(term.lang_tag, projected_header) ||
-        is_dynamic_annotation(term.data_type, projected_header)) {
-      return term;
-    }
-    if (term.data_type != "None" &&
-        !(term.data_type.starts_with("http://") || term.data_type.starts_with("https://"))) {
-      term.data_type = base_uri + term.data_type;
-    }
-  }
-
-  term.add_base_iri = term.term_type == "uri" || term.term_type == "iri" || term.term_type == "unsafeiri";
-
-  if (content[1] == "preformatted") {
-    term.map_type = CompiledTermMapType::Preformatted;
-  } else if (content[1] == "constant") {
-    term.map_type = CompiledTermMapType::Constant;
-    term.infer_datatype = is_literal && term.lang_tag == "None" && term.data_type == "None";
-  } else if (content[1] == "reference") {
-    term.map_type = CompiledTermMapType::Reference;
-    term.reference_index = projected_column_index(projected_header, term.term_map);
-    if (term.reference_index < 0) {
-      return CompiledTerm{};
-    }
-    term.infer_datatype = is_literal && term.lang_tag == "None" && term.data_type == "None";
-  } else if (content[1] == "template") {
-    term.map_type = CompiledTermMapType::Template;
-    if (!compile_template_parts(term.term_map, projected_header, term.parts)) {
-      return CompiledTerm{};
-    }
-  } else {
-    return term;
-  }
-
-  term.usable = true;
-  return term;
-}
-
-static bool term_needs_row_map(const std::vector<std::string>& content, const CompiledTerm& compiled) {
-  return content.size() > 1 && (content[1] == "function" || !compiled.usable);
-}
-
-static CompiledRuntimeTerm compile_runtime_term(const std::vector<std::string>& content,
-                                                const std::vector<std::string>& projected_header,
-                                                const std::string& base_uri) {
-  CompiledRuntimeTerm term;
-  term.content = content;
-  term.compiled = compile_term(term.content, projected_header, base_uri);
-  if (term.content.empty() || term.content[0] == "NULL") {
-    term.render_op = CompiledRuntimeTerm::RenderOp::Null;
-  } else if (term.content.size() > 1 && term.content[1] == "function") {
-    term.render_op = CompiledRuntimeTerm::RenderOp::Function;
-  } else if (term.content.size() > 1 && term.content[1] == "preformatted") {
-    term.render_op = CompiledRuntimeTerm::RenderOp::Preformatted;
-  } else if (term.compiled.usable) {
-    term.render_op = CompiledRuntimeTerm::RenderOp::Compiled;
-  } else {
-    term.render_op = CompiledRuntimeTerm::RenderOp::Fallback;
-  }
-  return term;
-}
-
-static bool can_cache_complex_term(const CompiledRuntimeTerm& term) {
-  if (term.render_op == CompiledRuntimeTerm::RenderOp::Preformatted) {
-    return true;
-  }
-  return term.render_op == CompiledRuntimeTerm::RenderOp::Compiled &&
-         term.compiled.usable &&
-         term.compiled.term_type != "blanknode";
-}
-
-static void append_complex_cache_key_part(std::string& key, std::string_view value) {
-  key += std::to_string(value.size());
-  key.push_back(':');
-  key.append(value);
-  key.push_back('|');
-}
-
-static std::string complex_term_cache_key(const std::string& base_uri,
-                                          const CompiledRuntimeTerm& term) {
-  std::string key;
-  key.reserve(base_uri.size() + term.content.size() * 24 + 32);
-  append_complex_cache_key_part(key, base_uri);
-  for (const std::string& part : term.content) {
-    append_complex_cache_key_part(key, part);
-  }
-  return key;
-}
-
-static void count_cacheable_complex_term(const std::string& base_uri,
-                                         const CompiledRuntimeTerm& term,
-                                         std::unordered_map<std::string, std::size_t>& counts) {
-  if (can_cache_complex_term(term)) {
-    counts[complex_term_cache_key(base_uri, term)]++;
-  }
-}
-
-static int assign_cacheable_complex_term(const std::string& base_uri,
-                                         const CompiledRuntimeTerm& term,
-                                         const std::unordered_map<std::string, std::size_t>& counts,
-                                         std::unordered_map<std::string, int>& ids,
-                                         std::vector<ComplexTermCacheEntry>& entries) {
-  if (!can_cache_complex_term(term)) {
-    return -1;
-  }
-
-  std::string key = complex_term_cache_key(base_uri, term);
-  auto count_it = counts.find(key);
-  if (count_it == counts.end() || count_it->second < 2) {
-    return -1;
-  }
-
-  auto id_it = ids.find(key);
-  if (id_it != ids.end()) {
-    return id_it->second;
-  }
-
-  const int id = static_cast<int>(entries.size());
-  ComplexTermCacheEntry entry;
-  entry.key = key;
-  entry.value.reserve(512);
-  entry.scratch.reserve(512);
-  entries.push_back(std::move(entry));
-  ids.emplace(std::move(key), id);
-  return id;
-}
-
-static void compile_complex_term_cache(CompiledComplexRenderPlan& plan,
-                                       const std::string& base_uri) {
-  std::unordered_map<std::string, std::size_t> counts;
-  counts.reserve(plan.has_graph ? 4 : 3);
-  count_cacheable_complex_term(base_uri, plan.subject, counts);
-  count_cacheable_complex_term(base_uri, plan.predicate, counts);
-  count_cacheable_complex_term(base_uri, plan.object, counts);
-  if (plan.has_graph) {
-    count_cacheable_complex_term(base_uri, plan.graph, counts);
-  }
-
-  std::unordered_map<std::string, int> ids;
-  ids.reserve(counts.size());
-  plan.term_cache_entries.reserve(counts.size());
-  plan.subject_cache_id = assign_cacheable_complex_term(base_uri, plan.subject, counts, ids, plan.term_cache_entries);
-  plan.predicate_cache_id = assign_cacheable_complex_term(base_uri, plan.predicate, counts, ids, plan.term_cache_entries);
-  plan.object_cache_id = assign_cacheable_complex_term(base_uri, plan.object, counts, ids, plan.term_cache_entries);
-  if (plan.has_graph) {
-    plan.graph_cache_id = assign_cacheable_complex_term(base_uri, plan.graph, counts, ids, plan.term_cache_entries);
-  }
-}
-
-static CompiledComplexRenderPlan compile_complex_render_plan(const std::vector<std::string>& joined_headers,
-                                                             const std::string& base_uri,
-                                                             const std::vector<std::string>& s_content,
-                                                             const std::vector<std::string>& p_content,
-                                                             const std::vector<std::string>& o_content,
-                                                             const std::vector<std::string>* g_content = nullptr) {
-  CompiledComplexRenderPlan plan;
-  plan.subject = compile_runtime_term(s_content, joined_headers, base_uri);
-  plan.predicate = compile_runtime_term(p_content, joined_headers, base_uri);
-  plan.object = compile_runtime_term(o_content, joined_headers, base_uri);
-  if (g_content != nullptr) {
-    plan.graph = compile_runtime_term(*g_content, joined_headers, base_uri);
-    plan.has_graph = true;
-  }
-
-  plan.has_object_condition = plan.object.content.size() > 5 && plan.object.content[5] != "None";
-  plan.needs_row_map = term_needs_row_map(plan.subject.content, plan.subject.compiled) ||
-                       term_needs_row_map(plan.predicate.content, plan.predicate.compiled) ||
-                       term_needs_row_map(plan.object.content, plan.object.compiled) ||
-                       plan.has_object_condition;
-  if (plan.has_graph) {
-    plan.needs_row_map = plan.needs_row_map || term_needs_row_map(plan.graph.content, plan.graph.compiled);
-  }
-  compile_complex_term_cache(plan, base_uri);
-  return plan;
-}
-
-static void render_compiled_term(const CompiledTerm& term,
-                                 const std::vector<std::string_view>& joined_row,
-                                 const std::string& base_uri,
-                                 std::string& out,
-                                 std::string& scratch) {
-  if (term.map_type == CompiledTermMapType::Preformatted) {
-    out = term.term_map;
-    return;
-  }
-
-  scratch.clear();
-  std::string_view rdf_term = term.term_map;
-  if (term.map_type == CompiledTermMapType::Reference) {
-    rdf_term = joined_row[term.reference_index];
-  } else if (term.map_type == CompiledTermMapType::Template) {
-    if (scratch.capacity() < term.term_map.size()) {
-      scratch.reserve(term.term_map.size());
-    }
-    for (const CompiledTemplatePart& part : term.parts) {
-      if (part.reference_index >= 0) {
-        const std::string_view value = joined_row[part.reference_index];
-        if (term.term_type == "uri") {
-          append_safe_iri(value, true, scratch);
-        } else if (term.term_type == "iri") {
-          append_safe_iri(value, false, scratch);
-        } else {
-          scratch += value;
-        }
-      } else {
-        scratch += part.literal;
-      }
-    }
-    rdf_term = scratch;
-  }
-
-  if ((term.map_type == CompiledTermMapType::Reference || term.map_type == CompiledTermMapType::Template) &&
-      term.add_base_iri &&
-      !(rdf_term.starts_with("http://") || rdf_term.starts_with("https://"))) {
-    if (rdf_term.data() != scratch.data()) {
-      scratch.assign(rdf_term);
-    }
-    scratch.insert(0, base_uri);
-    rdf_term = scratch;
-  }
-
-  const std::string datatype = term.infer_datatype ? infer_literal_datatype(rdf_term, term.lang_tag, term.data_type) : term.data_type;
-  handle_term_type_into(term.term_type, rdf_term, term.lang_tag, datatype, out);
 }
 
 static void initialize_row_map(std::unordered_map<std::string, std::string>& row,
@@ -472,9 +92,7 @@ static bool render_runtime_term(const CompiledRuntimeTerm& term,
 }
 
 static void reset_complex_term_cache(CompiledComplexRenderPlan& plan) {
-  for (ComplexTermCacheEntry& entry : plan.term_cache_entries) {
-    entry.computed = false;
-  }
+  reset_term_cache(plan.term_cache_entries);
 }
 
 static bool render_cached_runtime_term(const CompiledRuntimeTerm& term,
@@ -484,7 +102,7 @@ static bool render_cached_runtime_term(const CompiledRuntimeTerm& term,
                                        const std::string& base_uri,
                                        std::unordered_map<std::string, std::string>& row,
                                        int cache_id,
-                                       std::vector<ComplexTermCacheEntry>& cache_entries,
+                                       std::vector<TermCacheEntry>& cache_entries,
                                        std::string& out,
                                        std::string& scratch,
                                        std::string& function_value) {
@@ -492,7 +110,7 @@ static bool render_cached_runtime_term(const CompiledRuntimeTerm& term,
     return render_runtime_term(term, line_count, input_file_name, joined_row, base_uri, row, out, scratch, function_value);
   }
 
-  ComplexTermCacheEntry& entry = cache_entries[static_cast<std::size_t>(cache_id)];
+  TermCacheEntry& entry = cache_entries[static_cast<std::size_t>(cache_id)];
   if (!entry.computed) {
     entry.value.clear();
     if (!render_runtime_term(term, line_count, input_file_name, joined_row, base_uri, row, entry.value, entry.scratch, function_value)) {
