+++ src/flexrml/backend/executor/simple_executor.cpp
@@ -26,6 +26,9 @@
 
 #include "csv_row.h"
 #include "definitions.h"
+#include "simple_program.h"
+#include "term_cache.h"
+#include "term_program.h"
 #include "utils.h"
 #include "xxhash.h"
 
@@ -84,356 +87,6 @@ static void update_row_map(std::unordered_map<std::string, std::string>& row,
       it->second = projected_row[i];
     }
   }
-}
-
-enum class CompiledTermMapType {
-  Preformatted,
-  Constant,
-  Reference,
-  Template
-};
-
-struct CompiledTemplatePart {
-  std::string literal;
-  int reference_index = -1;
-};
-
-struct CompiledTerm {
-  bool usable = false;
-  CompiledTermMapType map_type = CompiledTermMapType::Constant;
-  std::string term_map;
-  std::string term_type;
-  std::string lang_tag = "None";
-  std::string data_type = "None";
-  std::vector<CompiledTemplatePart> parts;
-  int reference_index = -1;
-  bool infer_datatype = false;
-  bool add_base_iri = false;
-};
-
-static int projected_column_index(const std::vector<std::string>& projected_header,
-                                  std::string_view name) {
-  for (std::size_t i = 0; i < projected_header.size(); ++i) {
-    if (projected_header[i] == name) {
-      return static_cast<int>(i);
-    }
-  }
-  return -1;
-}
-
-static bool is_dynamic_annotation(const std::string& annotation,
-                                  const std::vector<std::string>& projected_header) {
-  return annotation.starts_with("==FUNC==") ||
-         annotation.find('{') != std::string::npos ||
-         projected_column_index(projected_header, annotation) >= 0;
-}
-
-static bool compile_template_parts(const std::string& term_map,
-                                   const std::vector<std::string>& projected_header,
-                                   std::vector<CompiledTemplatePart>& parts) {
-  std::string literal;
-  literal.reserve(term_map.size());
-
-  for (std::size_t i = 0; i < term_map.size();) {
-    if (term_map[i] == '\\' && i + 1 < term_map.size() &&
-        (term_map[i + 1] == '{' || term_map[i + 1] == '}')) {
-      literal.push_back(term_map[i + 1]);
-      i += 2;
-      continue;
-    }
-
-    if (term_map[i] != '{') {
-      literal.push_back(term_map[i]);
-      ++i;
-      continue;
-    }
-
-    const std::size_t placeholder_start = i;
-    ++i;
-    std::string reference_name;
-    bool closed = false;
-    for (; i < term_map.size(); ++i) {
-      if (term_map[i] == '\\' && i + 1 < term_map.size() &&
-          (term_map[i + 1] == '{' || term_map[i + 1] == '}')) {
-        reference_name.push_back(term_map[i + 1]);
-        ++i;
-      } else if (term_map[i] == '}') {
-        closed = true;
-        ++i;
-        break;
-      } else {
-        reference_name.push_back(term_map[i]);
-      }
-    }
-
-    if (!closed) {
-      literal.append(term_map, placeholder_start, std::string::npos);
-      break;
-    }
-
-    if (!literal.empty()) {
-      parts.push_back(CompiledTemplatePart{std::move(literal), -1});
-      literal.clear();
-    }
-
-    const int index = projected_column_index(projected_header, reference_name);
-    if (index < 0) {
-      return false;
-    }
-    parts.push_back(CompiledTemplatePart{"", index});
-  }
-
-  if (!literal.empty()) {
-    parts.push_back(CompiledTemplatePart{std::move(literal), -1});
-  }
-  return true;
-}
-
-static CompiledTerm compile_term(const std::vector<std::string>& content,
-                                 const std::vector<std::string>& projected_header,
-                                 const std::string& base_uri) {
-  CompiledTerm term;
-  if (content.size() < 3 || content[1] == "function") {
-    return term;
-  }
-
-  term.term_map = content[0];
-  term.term_type = content[2];
-  const bool is_literal = term.term_type == "literal";
-  if (is_literal) {
-    if (content.size() > 3) {
-      term.lang_tag = content[3];
-    }
-    if (content.size() > 4) {
-      term.data_type = content[4];
-    }
-    if (is_dynamic_annotation(term.lang_tag, projected_header) ||
-        is_dynamic_annotation(term.data_type, projected_header)) {
-      return term;
-    }
-    if (term.data_type != "None" &&
-        !(term.data_type.starts_with("http://") || term.data_type.starts_with("https://"))) {
-      term.data_type = base_uri + term.data_type;
-    }
-  }
-
-  term.add_base_iri = term.term_type == "uri" || term.term_type == "iri" || term.term_type == "unsafeiri";
-
-  if (content[1] == "preformatted") {
-    term.map_type = CompiledTermMapType::Preformatted;
-  } else if (content[1] == "constant") {
-    term.map_type = CompiledTermMapType::Constant;
-    term.infer_datatype = is_literal && term.lang_tag == "None" && term.data_type == "None";
-  } else if (content[1] == "reference") {
-    term.map_type = CompiledTermMapType::Reference;
-    term.reference_index = projected_column_index(projected_header, term.term_map);
-    if (term.reference_index < 0) {
-      return CompiledTerm{};
-    }
-    term.infer_datatype = is_literal && term.lang_tag == "None" && term.data_type == "None";
-  } else if (content[1] == "template") {
-    term.map_type = CompiledTermMapType::Template;
-    if (!compile_template_parts(term.term_map, projected_header, term.parts)) {
-      return CompiledTerm{};
-    }
-  } else {
-    return term;
-  }
-
-  term.usable = true;
-  return term;
-}
-
-static bool term_needs_row_map(const std::vector<std::string>& content, const CompiledTerm& compiled) {
-  return content.size() > 1 && (content[1] == "function" || !compiled.usable);
-}
-
-struct CompiledRuntimeTerm {
-  std::vector<std::string> content;
-  CompiledTerm compiled;
-  enum class RenderOp {
-    Null,
-    Function,
-    Preformatted,
-    Compiled,
-    Fallback
-  };
-  RenderOp render_op = RenderOp::Fallback;
-};
-
-struct CompiledSimplePlan {
-  CompiledRuntimeTerm subject;
-  CompiledRuntimeTerm predicate;
-  CompiledRuntimeTerm object;
-  CompiledRuntimeTerm graph;
-  bool has_graph = false;
-  bool needs_row_map = false;
-  bool function_called = false;
-  bool has_object_condition = false;
-};
-
-struct PreparedSimplePlan {
-  SimplePlan source;
-  std::vector<int> projected_indices;
-  std::vector<std::string> projected_header;
-  CompiledSimplePlan compiled;
-};
-
-std::vector<int> get_attribute_index(std::istream& file,
-                                     const std::vector<std::string>& header,
-                                     const std::vector<std::string>& projected_attributes);
-
-static CompiledRuntimeTerm compile_runtime_term(const std::vector<std::string>& content,
-                                                const std::vector<std::string>& projected_header,
-                                                const std::string& base_uri) {
-  CompiledRuntimeTerm term;
-  term.content = content;
-  term.compiled = compile_term(term.content, projected_header, base_uri);
-  if (term.content.empty() || term.content[0] == "NULL") {
-    term.render_op = CompiledRuntimeTerm::RenderOp::Null;
-  } else if (term.content.size() > 1 && term.content[1] == "function") {
-    term.render_op = CompiledRuntimeTerm::RenderOp::Function;
-  } else if (term.content.size() > 1 && term.content[1] == "preformatted") {
-    term.render_op = CompiledRuntimeTerm::RenderOp::Preformatted;
-  } else if (term.compiled.usable) {
-    term.render_op = CompiledRuntimeTerm::RenderOp::Compiled;
-  } else {
-    term.render_op = CompiledRuntimeTerm::RenderOp::Fallback;
-  }
-  return term;
-}
-
-static CompiledSimplePlan compile_simple_plan(const std::vector<std::string>& projected_header,
-                                              const std::string& base_uri,
-                                              const std::vector<std::string>& s_content,
-                                              const std::vector<std::string>& p_content,
-                                              const std::vector<std::string>& o_content,
-                                              const std::vector<std::string>* g_content = nullptr) {
-  CompiledSimplePlan plan;
-  plan.subject = compile_runtime_term(s_content, projected_header, base_uri);
-  plan.predicate = compile_runtime_term(p_content, projected_header, base_uri);
-  plan.object = compile_runtime_term(o_content, projected_header, base_uri);
-  if (g_content != nullptr) {
-    plan.graph = compile_runtime_term(*g_content, projected_header, base_uri);
-    plan.has_graph = true;
-  }
-
-  plan.has_object_condition = plan.object.content.size() > 5 && plan.object.content[5] != "None";
-  plan.needs_row_map = term_needs_row_map(plan.subject.content, plan.subject.compiled) ||
-                       term_needs_row_map(plan.predicate.content, plan.predicate.compiled) ||
-                       term_needs_row_map(plan.object.content, plan.object.compiled) ||
-                       plan.has_object_condition;
-  plan.function_called = plan.subject.render_op == CompiledRuntimeTerm::RenderOp::Function ||
-                         plan.predicate.render_op == CompiledRuntimeTerm::RenderOp::Function ||
-                         plan.object.render_op == CompiledRuntimeTerm::RenderOp::Function;
-  if (plan.has_graph) {
-    plan.needs_row_map = plan.needs_row_map || term_needs_row_map(plan.graph.content, plan.graph.compiled);
-    plan.function_called = plan.function_called || plan.graph.render_op == CompiledRuntimeTerm::RenderOp::Function;
-  }
-  return plan;
-}
-
-static bool prepare_simple_plan(const SimplePlan& source,
-                                std::istream& file,
-                                std::string& header_line,
-                                PreparedSimplePlan& prepared) {
-  if (!std::getline(file, header_line)) {
-    return false;
-  }
-
-  std::vector<std::string> header = split_csv_line(header_line, ',');
-  prepared.source = source;
-  if (!(source.projected_attributes.size() == 1 && source.projected_attributes[0] == "")) {
-    prepared.projected_indices = get_attribute_index(file, header, source.projected_attributes);
-  }
-
-  prepared.projected_header.clear();
-  prepared.projected_header.reserve(prepared.projected_indices.size());
-  for (int index : prepared.projected_indices) {
-    prepared.projected_header.push_back(header[index]);
-  }
-
-  prepared.compiled = compile_simple_plan(
-      prepared.projected_header,
-      source.base_uri,
-      source.s_content,
-      source.p_content,
-      source.o_content,
-      source.generate_graph ? &source.g_content : nullptr);
-  return true;
-}
-
-static void prepare_simple_plan_from_header(const SimplePlan& source,
-                                            const std::vector<std::string>& header,
-                                            PreparedSimplePlan& prepared) {
-  prepared.source = source;
-  if (!(source.projected_attributes.size() == 1 && source.projected_attributes[0] == "")) {
-    std::istringstream unused;
-    prepared.projected_indices = get_attribute_index(unused, header, source.projected_attributes);
-  }
-
-  prepared.projected_header.clear();
-  prepared.projected_header.reserve(prepared.projected_indices.size());
-  for (int index : prepared.projected_indices) {
-    prepared.projected_header.push_back(header[index]);
-  }
-
-  prepared.compiled = compile_simple_plan(
-      prepared.projected_header,
-      source.base_uri,
-      source.s_content,
-      source.p_content,
-      source.o_content,
-      source.generate_graph ? &source.g_content : nullptr);
-}
-
-static void render_compiled_term(const CompiledTerm& term,
-                                 const std::vector<std::string_view>& projected_row,
-                                 const std::string& base_uri,
-                                 std::string& out,
-                                 std::string& scratch) {
-  if (term.map_type == CompiledTermMapType::Preformatted) {
-    out = term.term_map;
-    return;
-  }
-
-  scratch.clear();
-  std::string_view rdf_term = term.term_map;
-  if (term.map_type == CompiledTermMapType::Reference) {
-    rdf_term = projected_row[term.reference_index];
-  } else if (term.map_type == CompiledTermMapType::Template) {
-    if (scratch.capacity() < term.term_map.size()) {
-      scratch.reserve(term.term_map.size());
-    }
-    for (const CompiledTemplatePart& part : term.parts) {
-      if (part.reference_index >= 0) {
-        const std::string_view value = projected_row[part.reference_index];
-        if (term.term_type == "uri") {
-          append_safe_iri(value, true, scratch);
-        } else if (term.term_type == "iri") {
-          append_safe_iri(value, false, scratch);
-        } else {
-          scratch += value;
-        }
-      } else {
-        scratch += part.literal;
-      }
-    }
-    rdf_term = scratch;
-  }
-
-  if ((term.map_type == CompiledTermMapType::Reference || term.map_type == CompiledTermMapType::Template) &&
-      term.add_base_iri &&
-      !(rdf_term.starts_with("http://") || rdf_term.starts_with("https://"))) {
-    if (rdf_term.data() != scratch.data()) {
-      scratch.assign(rdf_term);
-    }
-    scratch.insert(0, base_uri);
-    rdf_term = scratch;
-  }
-
-  const std::string datatype = term.infer_datatype ? infer_literal_datatype(rdf_term, term.lang_tag, term.data_type) : term.data_type;
-  handle_term_type_into(term.term_type, rdf_term, term.lang_tag, datatype, out);
 }
 
 static bool render_runtime_term(const CompiledRuntimeTerm& term,
@@ -775,120 +428,18 @@ static void split_project_current_line(SetupData& setup_data,
   project_row_views_from_strings(setup_data.projected_row, setup_data.projected_row_views);
 }
 
-struct FusedSimpleRuntime {
-  PreparedSimplePlan prepared;
-  std::vector<std::string> projected_row_storage;
-  std::vector<std::string_view> projected_row;
-  std::unordered_map<std::string, std::string> row;
-  int subject_cache_id = -1;
-  int predicate_cache_id = -1;
-  int object_cache_id = -1;
-  int graph_cache_id = -1;
-  std::string subject;
-  std::string predicate;
-  std::string object;
-  std::string graph;
-  std::string subject_scratch;
-  std::string predicate_scratch;
-  std::string object_scratch;
-  std::string graph_scratch;
-};
-
-struct FusedSimpleGroup {
-  std::vector<std::size_t> runtime_indices;
-  bool reuse_subject_predicate_object = false;
-};
-
-struct FusedTermCacheEntry {
-  std::string key;
-  std::string value;
-  std::string scratch;
-  bool computed = false;
-};
-
-struct FusedSimpleProgram {
-  std::vector<FusedSimpleRuntime> runtimes;
-  std::vector<FusedSimpleGroup> groups;
-  std::vector<FusedTermCacheEntry> term_cache_entries;
-};
-
-static bool can_cache_fused_term(const CompiledRuntimeTerm& term) {
-  if (term.render_op == CompiledRuntimeTerm::RenderOp::Preformatted) {
-    return true;
-  }
-  return term.render_op == CompiledRuntimeTerm::RenderOp::Compiled &&
-         term.compiled.usable &&
-         term.compiled.term_type != "blanknode";
-}
-
-static void append_key_part(std::string& key, std::string_view value) {
-  key += std::to_string(value.size());
-  key.push_back(':');
-  key.append(value);
-  key.push_back('|');
-}
-
-static std::string fused_term_cache_key(const std::string& base_uri,
-                                        const CompiledRuntimeTerm& term) {
-  std::string key;
-  key.reserve(base_uri.size() + term.content.size() * 24 + 32);
-  append_key_part(key, base_uri);
-  for (const std::string& part : term.content) {
-    append_key_part(key, part);
-  }
-  return key;
-}
-
-static void count_cacheable_fused_term(const std::string& base_uri,
-                                       const CompiledRuntimeTerm& term,
-                                       std::map<std::string, std::size_t>& counts) {
-  if (can_cache_fused_term(term)) {
-    counts[fused_term_cache_key(base_uri, term)]++;
-  }
-}
-
-static int assign_cacheable_fused_term(const std::string& base_uri,
-                                       const CompiledRuntimeTerm& term,
-                                       const std::map<std::string, std::size_t>& counts,
-                                       std::unordered_map<std::string, int>& ids,
-                                       std::vector<FusedTermCacheEntry>& entries) {
-  if (!can_cache_fused_term(term)) {
-    return -1;
-  }
-
-  std::string key = fused_term_cache_key(base_uri, term);
-  auto count_it = counts.find(key);
-  if (count_it == counts.end() || count_it->second < 2) {
-    return -1;
-  }
-
-  auto id_it = ids.find(key);
-  if (id_it != ids.end()) {
-    return id_it->second;
-  }
-
-  const int id = static_cast<int>(entries.size());
-  FusedTermCacheEntry entry;
-  entry.key = key;
-  entry.value.reserve(512);
-  entry.scratch.reserve(512);
-  entries.push_back(std::move(entry));
-  ids.emplace(std::move(key), id);
-  return id;
-}
-
 static bool append_cached_fast_term(const CompiledRuntimeTerm& term,
                                     const std::vector<std::string_view>& projected_row,
                                     const std::string& base_uri,
                                     int cache_id,
-                                    std::vector<FusedTermCacheEntry>& cache_entries,
+                                    std::vector<TermCacheEntry>& cache_entries,
                                     std::string& out,
                                     std::string& scratch) {
   if (cache_id < 0) {
     return append_fast_compiled_term(term, projected_row, base_uri, out, scratch);
   }
 
-  FusedTermCacheEntry& entry = cache_entries[static_cast<std::size_t>(cache_id)];
+  TermCacheEntry& entry = cache_entries[static_cast<std::size_t>(cache_id)];
   if (!entry.computed) {
     entry.value.clear();
     if (!append_fast_compiled_term(term, projected_row, base_uri, entry.value, entry.scratch)) {
@@ -907,7 +458,7 @@ static bool emit_cached_fast_statement(const CompiledSimplePlan& plan,
                                        int predicate_cache_id,
                                        int object_cache_id,
                                        int graph_cache_id,
-                                       std::vector<FusedTermCacheEntry>& cache_entries,
+                                       std::vector<TermCacheEntry>& cache_entries,
                                        std::string& out,
                                        std::string& subject_scratch,
                                        std::string& predicate_scratch,
@@ -940,150 +491,6 @@ static bool emit_cached_fast_statement(const CompiledSimplePlan& plan,
   return true;
 }
 
-static std::vector<FusedTermCacheEntry> compile_fused_term_cache(std::vector<FusedSimpleRuntime>& runtimes,
-                                                                 const std::vector<FusedSimpleGroup>& groups) {
-  std::vector<bool> skip_runtime(runtimes.size(), false);
-  for (std::size_t runtime_index = 0; runtime_index < runtimes.size(); ++runtime_index) {
-    if (runtimes[runtime_index].prepared.compiled.has_graph) {
-      skip_runtime[runtime_index] = true;
-    }
-  }
-  for (const FusedSimpleGroup& group : groups) {
-    if (!group.reuse_subject_predicate_object || group.runtime_indices.size() < 2) {
-      continue;
-    }
-    for (std::size_t runtime_index : group.runtime_indices) {
-      skip_runtime[runtime_index] = true;
-    }
-  }
-
-  std::map<std::string, std::size_t> counts;
-  for (std::size_t runtime_index = 0; runtime_index < runtimes.size(); ++runtime_index) {
-    if (skip_runtime[runtime_index]) {
-      continue;
-    }
-    const FusedSimpleRuntime& runtime = runtimes[runtime_index];
-    const CompiledSimplePlan& plan = runtime.prepared.compiled;
-    const std::string& base_uri = runtime.prepared.source.base_uri;
-    count_cacheable_fused_term(base_uri, plan.subject, counts);
-    count_cacheable_fused_term(base_uri, plan.predicate, counts);
-    count_cacheable_fused_term(base_uri, plan.object, counts);
-    if (plan.has_graph) {
-      count_cacheable_fused_term(base_uri, plan.graph, counts);
-    }
-  }
-
-  std::unordered_map<std::string, int> ids;
-  std::vector<FusedTermCacheEntry> entries;
-  ids.reserve(counts.size());
-  entries.reserve(counts.size());
-  for (std::size_t runtime_index = 0; runtime_index < runtimes.size(); ++runtime_index) {
-    if (skip_runtime[runtime_index]) {
-      continue;
-    }
-    FusedSimpleRuntime& runtime = runtimes[runtime_index];
-    const CompiledSimplePlan& plan = runtime.prepared.compiled;
-    const std::string& base_uri = runtime.prepared.source.base_uri;
-    runtime.subject_cache_id = assign_cacheable_fused_term(base_uri, plan.subject, counts, ids, entries);
-    runtime.predicate_cache_id = assign_cacheable_fused_term(base_uri, plan.predicate, counts, ids, entries);
-    runtime.object_cache_id = assign_cacheable_fused_term(base_uri, plan.object, counts, ids, entries);
-    if (plan.has_graph) {
-      runtime.graph_cache_id = assign_cacheable_fused_term(base_uri, plan.graph, counts, ids, entries);
-    }
-  }
-  return entries;
-}
-
-static void reset_fused_term_cache(FusedSimpleProgram& program) {
-  for (FusedTermCacheEntry& entry : program.term_cache_entries) {
-    entry.computed = false;
-  }
-}
-
-static bool has_cached_fused_term(const FusedSimpleRuntime& runtime) {
-  return runtime.subject_cache_id >= 0 ||
-         runtime.predicate_cache_id >= 0 ||
-         runtime.object_cache_id >= 0 ||
-         runtime.graph_cache_id >= 0;
-}
-
-static void initialize_fused_runtime(FusedSimpleRuntime& runtime) {
-  runtime.projected_row_storage.reserve(runtime.prepared.projected_indices.size());
-  runtime.projected_row.reserve(runtime.prepared.projected_indices.size());
-  runtime.subject.reserve(512);
-  runtime.predicate.reserve(512);
-  runtime.object.reserve(512);
-  runtime.graph.reserve(512);
-  runtime.subject_scratch.reserve(512);
-  runtime.predicate_scratch.reserve(512);
-  runtime.object_scratch.reserve(512);
-  runtime.graph_scratch.reserve(512);
-  if (runtime.prepared.compiled.needs_row_map) {
-    initialize_row_map(runtime.row, runtime.prepared.projected_header);
-  }
-}
-
-static bool can_reuse_subject_predicate_object(const FusedSimpleRuntime& runtime) {
-  const CompiledSimplePlan& plan = runtime.prepared.compiled;
-  return plan.has_graph &&
-         !plan.needs_row_map &&
-         !plan.function_called &&
-         !plan.has_object_condition;
-}
-
-static bool has_same_subject_predicate_object_group(const FusedSimpleRuntime& left,
-                                                    const FusedSimpleRuntime& right) {
-  return left.prepared.source.base_uri == right.prepared.source.base_uri &&
-         left.prepared.source.projected_attributes == right.prepared.source.projected_attributes &&
-         left.prepared.source.s_content == right.prepared.source.s_content &&
-         left.prepared.source.p_content == right.prepared.source.p_content &&
-         left.prepared.source.o_content == right.prepared.source.o_content;
-}
-
-static std::vector<FusedSimpleGroup> build_fused_simple_groups(const std::vector<FusedSimpleRuntime>& runtimes) {
-  std::vector<FusedSimpleGroup> groups;
-  for (std::size_t i = 0; i < runtimes.size(); ++i) {
-    if (!can_reuse_subject_predicate_object(runtimes[i])) {
-      groups.push_back(FusedSimpleGroup{{i}, false});
-      continue;
-    }
-
-    bool added = false;
-    for (auto& group : groups) {
-      if (!group.reuse_subject_predicate_object) {
-        continue;
-      }
-      const FusedSimpleRuntime& representative = runtimes[group.runtime_indices.front()];
-      if (has_same_subject_predicate_object_group(representative, runtimes[i])) {
-        group.runtime_indices.push_back(i);
-        added = true;
-        break;
-      }
-    }
-
-    if (!added) {
-      groups.push_back(FusedSimpleGroup{{i}, true});
-    }
-  }
-  return groups;
-}
-
-static FusedSimpleProgram compile_fused_simple_program(const std::vector<SimplePlan>& plans,
-                                                       const std::vector<std::string>& header) {
-  FusedSimpleProgram program;
-  program.runtimes.reserve(plans.size());
-  for (const auto& plan : plans) {
-    FusedSimpleRuntime runtime;
-    prepare_simple_plan_from_header(plan, header, runtime.prepared);
-    initialize_fused_runtime(runtime);
-    program.runtimes.push_back(std::move(runtime));
-  }
-
-  program.groups = build_fused_simple_groups(program.runtimes);
-  program.term_cache_entries = compile_fused_term_cache(program.runtimes, program.groups);
-  return program;
-}
-
 static bool is_single_constant_statement(const SimplePlan& plan) {
   if (!plan.generate_graph) {
     return (plan.s_content[1] == "constant" && plan.p_content[1] == "constant" && plan.o_content[1] == "constant") ||
@@ -1091,31 +498,6 @@ static bool is_single_constant_statement(const SimplePlan& plan) {
   }
   return (plan.s_content[1] == "constant" && plan.p_content[1] == "constant" && plan.o_content[1] == "constant" && plan.g_content[1] == "constant") ||
          (plan.s_content[1] == "preformatted" && plan.p_content[1] == "preformatted" && plan.o_content[1] == "preformatted" && plan.g_content[1] == "preformatted");
-}
-
-///////////////////////////////////////////////////////////////7
-/// HELPER FUNCTIONS
-//////////////////////////////////////////////////////////////
-std::vector<int> get_attribute_index(std::istream& file, const std::vector<std::string>& header, const std::vector<std::string>& projected_attributes) {
-  // Get indices for projected attributes
-  std::vector<int> projected_indices;
-
-  for (const auto& attr : projected_attributes) {
-    if (attr == "") {
-      continue;
-    }
-
-    auto it = std::find_if(header.begin(), header.end(), [&](std::string_view field) {
-      return field == attr;
-    });
-    if (it != header.end()) {
-      projected_indices.push_back(std::distance(header.begin(), it));
-    } else {
-      throw std::runtime_error("Attribute not found: '" + attr + "'");
-    }
-  }
-
-  return projected_indices;
 }
 
 // Open File or from map
@@ -1778,7 +1160,7 @@ size_t execute_fused_simple_plans(const std::vector<SimplePlan>& plans,
   FusedSimpleProgram program = compile_fused_simple_program(plans, header);
   std::vector<FusedSimpleRuntime>& runtimes = program.runtimes;
   const std::vector<FusedSimpleGroup>& groups = program.groups;
-  std::vector<FusedTermCacheEntry>& term_cache_entries = program.term_cache_entries;
+  std::vector<TermCacheEntry>& term_cache_entries = program.term_cache_entries;
 
   int line_count = 0;
   while (std::getline(*file, setup_data.line)) {
