+++ src/flexrml/backend/executor/simple_executor.h
@@ -6,22 +6,12 @@
 #include <unordered_set>
 #include <unordered_map>
 
+#include "simple_plan.h"
+
 class OutputChunkWriter {
  public:
   virtual ~OutputChunkWriter() = default;
   virtual void write(std::string chunk) = 0;
-};
-
-struct SimplePlan {
-  std::filesystem::path output_file_name;
-  std::string base_uri;
-  std::string input_file_name;
-  std::vector<std::string> projected_attributes;
-  std::vector<std::string> s_content;
-  std::vector<std::string> p_content;
-  std::vector<std::string> o_content;
-  std::vector<std::string> g_content;
-  bool generate_graph = false;
 };
 
 SimplePlan parse_simple_plan(const std::string& information);
