+++ src/flexrml/cli.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "options.h"
+
+Options parse_args(int argc, char** argv);
