+++ src/flexrml/pipeline.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <string>
+
+#include "options.h"
+
+std::string generate_plan(const Options& options);
+std::string execute_mapping(const Options& options);
